Server/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
psycopg2==2.9.11
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
djangorestframework-simplejwt
pandas